// src/app/models/leaveModel.ts
export interface LeapApprovalStatus {
  approval_type: string
}

export interface LeapClientLeave {
  leave_name: string
}

export interface LeapCustomer {
  name: string
}
// ...
export interface CustomerLeavePendingCount {
  leaveTypeId: number
  leaveType: string
  leaveAllotedCount: number
  totalAppliedLeaveDays: number
  leaveBalance: number
  color_code:any
}
// ...
export interface EmpLeave {
  id: number
  created_at: string
  customer_id: number
  client_id: number
  leave_type: number
  from_date: string
  to_date: string
  total_days: number
  approved_by_id?: number
  leave_status: number
  leave_reason: string
  approve_disapprove_remark?: string
  attachments?: string
  branch_id: number
  updated_at: string
  isAssigned: boolean
  duration: string
  leap_approval_status: LeapApprovalStatus
  leap_client_leave: LeapClientLeave
  leap_customer: LeapCustomer
}